push eka

+++ routes/web.php
@@ -29,48 +29,48 @@ Route::get('forgot-password', [AuthController::class, 'showLinkRequestForm'])->n
 Route::post('forgot-password', [AuthController::class, 'sendResetLink'])->name('password.update');
 
 
-Route::get('/', [KriteriaController::class, 'home'])->name('home');
-Route::get('/user', [UserController::class, 'index'])->name('user');
-Route::post('/user/tambah', [UserController::class, 'tambahUser']);
-Route::get('/user/edit/{kode}', [userController::class, 'editUser']);
-Route::post('/user/edit', [userController::class, 'updateUser']);
-Route::delete('/user/{kode}', [userController::class, 'destroy']);
+Route::get('/', [KriteriaController::class, 'home'])->name('home')->middleware('auth');
+Route::get('/user', [UserController::class, 'index'])->name('user')->middleware('auth');
+Route::post('/user/tambah', [UserController::class, 'tambahUser'])->middleware('auth');
+Route::get('/user/edit/{kode}', [userController::class, 'editUser'])->middleware('auth');
+Route::post('/user/edit', [userController::class, 'updateUser'])->middleware('auth');
+Route::delete('/user/{kode}', [userController::class, 'destroy'])->middleware('auth');
 
-Route::get('/produk', [ProdukController::class, 'index'])->name('produk');
-Route::post('/produk/tambah', [ProdukController::class, 'tambahProduk']);
-Route::get('/produk/edit/{kode}', [ProdukController::class, 'editProduk']);
-Route::post('/produk/edit', [ProdukController::class, 'updateProduk']);
-Route::delete('/produk/{kode}', [ProdukController::class, 'destroy']);
+Route::get('/produk', [ProdukController::class, 'index'])->name('produk')->middleware('auth');
+Route::post('/produk/tambah', [ProdukController::class, 'tambahProduk'])->middleware('auth');
+Route::get('/produk/edit/{kode}', [ProdukController::class, 'editProduk'])->middleware('auth');
+Route::post('/produk/edit', [ProdukController::class, 'updateProduk'])->middleware('auth');
+Route::delete('/produk/{kode}', [ProdukController::class, 'destroy'])->middleware('auth');
 
-Route::get('/kriteria', [KriteriaController::class, 'index'])->name('kriteria');
-Route::post('/kriteria/tambah', [KriteriaController::class, 'tambahKriteria']);
-Route::get('/kriteria/edit/{kode}', [KriteriaController::class, 'editkriteria']);
-Route::post('/kriteria/edit', [KriteriaController::class, 'updateKriteria']);
-Route::delete('/kriteria/{kode}', [KriteriaController::class, 'destroy']);
+Route::get('/kriteria', [KriteriaController::class, 'index'])->name('kriteria')->middleware('auth');
+Route::post('/kriteria/tambah', [KriteriaController::class, 'tambahKriteria'])->middleware('auth');
+Route::get('/kriteria/edit/{kode}', [KriteriaController::class, 'editkriteria'])->middleware('auth');
+Route::post('/kriteria/edit', [KriteriaController::class, 'updateKriteria'])->middleware('auth');
+Route::delete('/kriteria/{kode}', [KriteriaController::class, 'destroy'])->middleware('auth');
 
-Route::get('/alternatif', [AlternativeController::class, 'index'])->name('alternatif');
-Route::post('/alternatif/tambah', [AlternativeController::class, 'tambahAlternatif']);
-Route::get('/alternatif/edit/{kode}', [AlternativeController::class, 'editAlternatif']);
-Route::post('/alternatif/edit', [ALternativeController::class, 'updateAlternatif']);
-Route::delete('/alternatif/{kode}', [AlternativeController::class, 'destroy']);
+Route::get('/alternatif', [AlternativeController::class, 'index'])->name('alternatif')->middleware('auth');
+Route::post('/alternatif/tambah', [AlternativeController::class, 'tambahAlternatif'])->middleware('auth');
+Route::get('/alternatif/edit/{kode}', [AlternativeController::class, 'editAlternatif'])->middleware('auth');
+Route::post('/alternatif/edit', [ALternativeController::class, 'updateAlternatif'])->middleware('auth');
+Route::delete('/alternatif/{kode}', [AlternativeController::class, 'destroy'])->middleware('auth');
 
-Route::get('/surat/pengajuan', [SuratController::class, 'index'])->name('surat');
-Route::get('/validasi', [SuratController::class, 'indexValidasi'])->name('suratValidasi');
-Route::post('/pengajuan/tambah', [SuratController::class, 'submitRequest']);
-Route::get('/pengajuan/edit/{kode}', [SuratController::class, 'editPengajuan']);
-Route::post('/pengajuan/edit', [SuratController::class, 'updatePengajuan']);
-Route::delete('/pengajuan/{kode}', [SuratController::class, 'destroy']);
-Route::get('/partial/modal/decline/{kode}', [SuratController::class, 'modalDecline'])->name('modalDecline');
-Route::post('/save/notes/{kode}', [SuratController::class, 'submitModalDecline'])->name('submitDecline');
-Route::post('/validate/pengajuan/{kode}', [SuratController::class, 'validasiPengajuan'])->name('validateSPK');
+Route::get('/surat/pengajuan', [SuratController::class, 'index'])->name('surat')->middleware('auth');
+Route::get('/validasi', [SuratController::class, 'indexValidasi'])->name('suratValidasi')->middleware('auth');
+Route::post('/pengajuan/tambah', [SuratController::class, 'submitRequest'])->middleware('auth');
+Route::get('/pengajuan/edit/{kode}', [SuratController::class, 'editPengajuan'])->middleware('auth');
+Route::post('/pengajuan/edit', [SuratController::class, 'updatePengajuan'])->middleware('auth');
+Route::delete('/pengajuan/{kode}', [SuratController::class, 'destroy'])->middleware('auth');
+Route::get('/partial/modal/decline/{kode}', [SuratController::class, 'modalDecline'])->name('modalDecline')->middleware('auth');
+Route::post('/save/notes/{kode}', [SuratController::class, 'submitModalDecline'])->name('submitDecline')->middleware('auth');
+Route::post('/validate/pengajuan/{kode}', [SuratController::class, 'validasiPengajuan'])->name('validateSPK')->middleware('auth');
 
-// Route::get('/penilaian', [PenilaianController::class, 'index']);
-Route::post('/penilaian/tambah', [PenilaianController::class, 'tambahPenilaian']);
-Route::get('/penilaian/edit/{alternatif}', [PenilaianController::class, 'editPenilaian']);
-Route::post('/penilaian/Edit', [PenilaianController::class, 'updatePenilaian']);
+// Route::get('/penilaian', [PenilaianController::class, 'index'])->middleware('auth');
+Route::post('/penilaian/tambah', [PenilaianController::class, 'tambahPenilaian'])->middleware('auth');
+Route::get('/penilaian/edit/{alternatif}', [PenilaianController::class, 'editPenilaian'])->middleware('auth');
+Route::post('/penilaian/Edit', [PenilaianController::class, 'updatePenilaian'])->middleware('auth');
 
-Route::get('/perhitungan', [PenilaianController::class, 'indexPerhitungan']);
+Route::get('/perhitungan', [PenilaianController::class, 'indexPerhitungan'])->middleware('auth');
 
-Route::get('/hasil', [PenilaianController::class, 'hasil']);
+Route::get('/hasil', [PenilaianController::class, 'hasil'])->middleware('auth');
 
-Route::get('/print', [PenilaianController::class, 'indexPrint'])->name('print');
+Route::get('/print', [PenilaianController::class, 'indexPrint'])->name('print')->middleware('auth');
